azure-lab2/5-durable-mapreduce/requirements.txt
azure-functions==1.24.0
azure-functions-durable==1.4.0
azure-storage-blob==12.28.0
